refactor: refactor setup template for vue3
